+++ src/Modules/RELAY_MODULE/Relay.php
@@ -6,7 +6,6 @@ use function Amp\asyncCall;
 
 use Amp\Loop;
 use Generator;
-use Nadybot\Core\Modules\ALTS\AltsController;
 use Nadybot\Core\{
 	Attributes as NCA,
 	ConfigFile,
@@ -50,12 +49,6 @@ class Relay implements MessageReceiver {
 
 	#[NCA\Inject]
 	public StatsController $statsController;
-
-	#[NCA\Inject]
-	public RelayController $relayController;
-
-	#[NCA\Inject]
-	public AltsController $altsController;
 
 	#[NCA\Logger]
 	public LoggerWrapper $logger;
@@ -108,23 +101,7 @@ class Relay implements MessageReceiver {
 
 	/** @return array<string,array<string,OnlinePlayer>> */
 	public function getOnlineList(): array {
-		if (!$this->relayController->relayUseLocalMain) {
-			return $this->onlineChars;
-		}
-		$result = [];
-		foreach ($this->onlineChars as $relay => $charLists) {
-			$result[$relay] = [];
-			foreach ($charLists as $name => $char) {
-				$newChar = clone $char;
-				$main = $this->altsController->getMainOf($char->pmain);
-				if ($main === $char->pmain) {
-					$main = $this->altsController->getMainOf($char->name);
-				}
-				$newChar->pmain = $main;
-				$result[$relay][$name] = $newChar;
-			}
-		}
-		return $result;
+		return $this->onlineChars;
 	}
 
 	public function clearOnline(string $where): void {