+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `!wish`-command to manage wishlists
 - Add 2 new options for raid rewards: `max_raid_reward_height` to limit the amount of points `!raid reward` and `!raid punish` can give/take (does not apply to pre-defined raid-rewards) and `raid_reward_predefined_only` to limit `!raid reward` and `!raid punish` to names of raid rewards instead of arbitrary points.
 - Add `!sites ql <ql>` to search for unplanted sites able to hold towers of a given QL.
+- Sharing online-lists via nadynative will now also share each character's main. This allows two new grouping options for relay online lists: by main (player) and by org, then main (org/player). This only works properly if all Nadybots are running the same version.
 
 ### Changed
 
@@ -50,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The console history works 100% now, only ctrl+r-search is now broken.
 - The bot now correctly sets the name of the character who added someone to the tracker.
 - The timer you can enable when a tower field goes down sometimes couldn't be created, because a timer with the same name was already there. Any timer with the same name will now be deleted before creating a new one.
+- Not grouping relay online lists gives a proper "Alliance"-group again
 
 ## [6.0.5] - 2022-06-17